requirements.txt
streamlit
langchain
langchain-community
openai
pypdf
tiktoken
PyPDF2
python-dotenv
langchain-openai